% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{get_url}
\alias{get_url}
\title{Title}
\usage{
get_url(id = "url")
}
\arguments{
\item{id}{the id of the module, not really needed for this use case}
}
\value{
the url of the current shiny page, it will only execute it once reducing waiting times
}
\description{
Title
}
\examples{
\dontrun{
get_url()
}
}
