% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedInShare.R
\name{LinkedInShare}
\alias{LinkedInShare}
\title{Creates the Shiny server module for twitter using moduleServer}
\usage{
LinkedInShare(id, url = NULL)
}
\arguments{
\item{id}{The module id must match the UI}

\item{url}{a valid url to be shared on LinkedIn}
}
\description{
Creates the Shiny server module for twitter using moduleServer
}
\examples{
\dontrun{
LinkedInShare("LinkedInShare")
}
}
