% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Share.R
\name{Share}
\alias{Share}
\title{Creates the Shiny server module for socual networks using moduleServer}
\usage{
Share(id, text = "This is my shiny app!", hashtags = "rstats,r", url = NULL)
}
\description{
Creates the Shiny server module for socual networks using moduleServer
}
