% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_message_linkedin.R
\name{share_message_linkedin}
\alias{share_message_linkedin}
\title{Creates the twitter message to be shared}
\usage{
share_message_linkedin(url)
}
\arguments{
\item{url}{a valid url to be shared on LinkedIn}
}
\value{
a string with the concatenation of the arguments plus the required skeleton.
}
\description{
Creates the twitter message to be shared
}
\examples{
share_message_twitter(text = "see", url = "how", "it works")
}
