% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_message_twitter.R
\name{share_message_twitter}
\alias{share_message_twitter}
\title{Creates the twitter message to be shared}
\usage{
share_message_twitter(text, url, hashtags)
}
\arguments{
\item{text}{a string which will begin the tweet}

\item{url}{a valid url to be shared}

\item{hashtags}{the end of the tweet separated by commas}
}
\value{
a string with the concatenation of the arguments plus the required skeleton.
}
\description{
Creates the twitter message to be shared
}
\examples{
share_message_twitter(text = "see", url = "how","it works")
}
