% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitterShare.R
\name{twitterShare}
\alias{twitterShare}
\title{Creates the Shiny server module for twitter using moduleServer}
\usage{
twitterShare(
  id,
  text = "This is my shiny app!",
  hashtags = "rstats,r",
  url = NULL
)
}
\arguments{
\item{id}{The module id must match the UI}

\item{text}{a string which will begin the tweet}

\item{hashtags}{the end of the tweet separated by commas}

\item{url}{a valid url to be shared on twitter}
}
\description{
Creates the Shiny server module for twitter using moduleServer
}
\examples{
\dontrun{
twitterShare("TwitterShare")
}
}
