% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitterShare.R
\name{twitterShareUI}
\alias{twitterShareUI}
\title{Creates the Shiny UI for twitter using uiOutput}
\usage{
twitterShareUI(id, inline = TRUE, container = if (inline) span else div, ...)
}
\arguments{
\item{id}{The module id must match the Server}

\item{inline}{use an inline (\code{span()}) or block container (\code{div()})
for the output}

\item{container}{a function to generate an HTML element to contain the text}

\item{...}{Other arguments to pass to the container tag function. This is
useful for providing additional classes for the tag.}
}
\value{
An HTML output element that can be included in a panel
}
\description{
Render a reactive output variable as HTML within an application page. The
text will be included within an HTML \code{div} tag, and is presumed to
contain HTML content which should not be escaped.
}
\details{
\code{uiOutput} is intended to be used with \code{renderUI} on the server
side. It is currently just an alias for \code{htmlOutput}.
}
\examples{
\dontrun{
twitterShareUI("TwitterShare")
}
}
